refactor: Compose elements for potential reusability and set stripe instance naming, docs
